pyproject.toml
[tool.poetry]
name = "flask-sketch"
version = "0.2.2"
description = "A CLI for autogenerate folder structure for Flask applications"
authors = ["Eric Souza <ericsouza0801@gmail.com>"]
license = "MIT License"

[tool.poetry.dependencies]
python = ">=3.7,<4.0"
pyfiglet = "^0.8.post1"
pyinquirer = "^1.0.3"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
flake8 = "^3.8.3"
flask-restx = "^0.2.0"
flask = "^1.1.2"
flask-sqlalchemy = "^2.4.4"
flask-praetorian = "^1.0.0"
flask-admin = "^1.5.6"
flask-basicauth = "^0.2.0"
python-dotenv = "^0.14.0"
flask-migrate = "^2.5.3"
flask-limiter = "^1.3.1"
flask-caching = "^1.9.0"
pyctuator = "^0.12"
psutil = "^5.7.2"

[tool.poetry.scripts]
flask-sketch = "flask_sketch.__main__:flask_sketch"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
